+++ modules/canbus/vehicle/lexus/protocol/parking_brake_cmd_124.cc
@@ -63,7 +63,7 @@ Parkingbrakecmd124* Parkingbrakecmd124::set_ignore_overrides(
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Parkingbrakecmd124::set_p_ignore_overrides(uint8_t* data,
                                                 bool ignore_overrides) {
-  int x = ignore_overrides;
+  uint8_t x = ignore_overrides;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 1, 1);
@@ -78,7 +78,7 @@ Parkingbrakecmd124* Parkingbrakecmd124::set_enable(bool enable) {
 // 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 0, 'type': 'bool',
 // 'order': 'motorola', 'physical_unit': ''}
 void Parkingbrakecmd124::set_p_enable(uint8_t* data, bool enable) {
-  int x = enable;
+  uint8_t x = enable;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 0, 1);
@@ -95,7 +95,7 @@ Parkingbrakecmd124* Parkingbrakecmd124::set_clear_override(
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Parkingbrakecmd124::set_p_clear_override(uint8_t* data,
                                               bool clear_override) {
-  int x = clear_override;
+  uint8_t x = clear_override;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 2, 1);
@@ -112,7 +112,7 @@ Parkingbrakecmd124* Parkingbrakecmd124::set_parking_brake_cmd(
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Parkingbrakecmd124::set_p_parking_brake_cmd(uint8_t* data,
                                                  bool parking_brake_cmd) {
-  int x = parking_brake_cmd;
+  uint8_t x = parking_brake_cmd;
 
   Byte to_set(data + 1);
   to_set.set_value(x, 0, 1);
@@ -127,7 +127,7 @@ Parkingbrakecmd124* Parkingbrakecmd124::set_clear_faults(bool clear_faults) {
 // 'len': 1, 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 3,
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Parkingbrakecmd124::set_p_clear_faults(uint8_t* data, bool clear_faults) {
-  int x = clear_faults;
+  uint8_t x = clear_faults;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 3, 1);
